Rename file with game objects logic
