+++ src/game-objects.js
